ClockSkew was added to middleware.

+++ Core/Utilities/Security/JWT/TokenOptions.cs
@@ -4,7 +4,7 @@
     {
         public string Audience { get; set; }
         public string Issuer { get; set; }
-        public int AccessTokenExpiration { get; set; }
+        public double AccessTokenExpiration { get; set; }
         public string SecurityKey { get; set; }
     }
 }
